requirements.txt
pyngrok==7.2.0
streamlit==1.38.0
PyPDF2==3.0.1
langchain==0.2.15
langchain-core==0.2.37
langchain-community==0.2.15
spacy==3.7.6
faiss-cpu==1.8.0.post1
python-dotenv==1.0.1
langchain-anthropic==0.1.23
langchain-openai==0.1.23
litellm==1.44.14
litellm[proxy]